We can now use the rsyn in order to sync the data

+++ mainframer.sh
@@ -16,6 +16,7 @@ REMOTE_BUILD_MACHINE=$(awk -F "=" '/remote_build.machine/ {print $2}' "$PROJECT_
 LOCAL_GZIP_LEVEL=$(awk -F "=" '/remote_build.local_gzip_level/ {print $2}' "$PROJECT_DIR/local.properties")
 REMOTE_GZIP_LEVEL=$(awk -F "=" '/remote_build.remote_gzip_level/ {print $2}' "$PROJECT_DIR/local.properties")
 REMOTE_SSH_PORT=$(awk -F "=" '/remote_build.ssh_port/ {print $2}' "$PROJECT_DIR/local.properties")
+USE_RSYNC=$(awk -F "=" '/build.use_rsync/ {print $2}' "$PROJECT_DIR/local.properties")
 
 if [ -z "$LOCAL_GZIP_LEVEL" ]; then
 	LOCAL_GZIP_LEVEL=1
@@ -32,6 +33,10 @@ fi
 if [ -z "$REMOTE_BUILD_MACHINE" ]; then
 	echo "Please specify remote build machine in local.properties"
 	exit 1
+fi
+
+if [ -z "$USE_RSYNC" ]; then
+	USE_RSYNC=false
 fi
 
 BUILD_COMMAND="$@"
@@ -67,6 +72,21 @@ LOCAL_ARCHIVE_COMMAND="tar \
 --exclude='*/build' \
 ."
 
+RSYNC_TO_REMOTE="rsync -av --delete \
+--exclude='**/.git/' \
+--exclude='**/local.properties' \
+--exclude='.idea' \
+--exclude='.gradle' \
+--exclude='**/build' \
+-e 'ssh -p $REMOTE_SSH_PORT' ./ $REMOTE_BUILD_MACHINE:~/$PROJECT_DIR_NAME/"
+
+RSYNC_FROM_REMOTE="rsync --delete -av --partial \
+--exclude='**/.git/' \
+--exclude='**/local.properties' \
+--exclude='.idea' \
+--exclude='.gradle' \
+-e 'ssh -p $REMOTE_SSH_PORT' $REMOTE_BUILD_MACHINE:~/$PROJECT_DIR_NAME/ ./"
+
 if [ $LOCAL_GZIP_LEVEL = "0" ]; then
 	LOCAL_ARCHIVE_COMMAND+=" > build/project_for_remote_build.tar"
 	REMOTE_UNARCHIVE_COMMAND="tar -xf project_for_remote_build.tar -C $PROJECT_DIR_NAME"
@@ -75,7 +95,9 @@ else
 	REMOTE_UNARCHIVE_COMMAND="\$REMOTE_ARCHIVER -d < project_for_remote_build.tar | tar -xf - -C $PROJECT_DIR_NAME"
 fi
 
-eval $LOCAL_ARCHIVE_COMMAND
+if [ $USE_RSYNC == false ]; then
+	eval $LOCAL_ARCHIVE_COMMAND
+fi
 popd
 
 # Prepare remote archive and local unarchive commands.
@@ -94,32 +116,52 @@ else
 fi
 
 # Transfer archive to remote machine.
-scp -P "$REMOTE_SSH_PORT" "$PROJECT_DIR/build/project_for_remote_build.tar" $REMOTE_BUILD_MACHINE:~/
+
+if [ $USE_RSYNC == false ]; then
+	scp -P "$REMOTE_SSH_PORT" "$PROJECT_DIR/build/project_for_remote_build.tar" $REMOTE_BUILD_MACHINE:~/
+else
+	eval $RSYNC_TO_REMOTE
+fi
 
 # Build project on a remote machine and then archive it.
-ssh $REMOTE_BUILD_MACHINE -p "$REMOTE_SSH_PORT" \
-"set -xe && \
-export ANDROID_HOME=/home/kokeroulis/Android/Sdk \
-export REMOTE_ARCHIVER=\`eval \"$DETECT_ARCHIVER\"\` && \
-cd ~ && \
-mkdir -p $PROJECT_DIR_NAME && \
-rm -rf $PROJECT_DIR_NAME/build/remotely_built_project.tar $PROJECT_DIR_NAME/*/src && \
-$REMOTE_UNARCHIVE_COMMAND && \
-cd $PROJECT_DIR_NAME && \
-$BUILD_COMMAND && \
-$REMOTE_ARCHIVE_COMMAND"
+if [ $USE_RSYNC == false ]; then
+	ssh $REMOTE_BUILD_MACHINE -p "$REMOTE_SSH_PORT" \
+	"set -xe && \
+	export ANDROID_HOME=/home/kokeroulis/Android/Sdk \
+	export REMOTE_ARCHIVER=\`eval \"$DETECT_ARCHIVER\"\` && \
+	cd ~ && \
+	mkdir -p $PROJECT_DIR_NAME && \
+	rm -rf $PROJECT_DIR_NAME/build/remotely_built_project.tar $PROJECT_DIR_NAME/*/src && \
+	$REMOTE_UNARCHIVE_COMMAND && \
+	cd $PROJECT_DIR_NAME && \
+	$BUILD_COMMAND && \
+	$REMOTE_ARCHIVE_COMMAND"
+else
+	ssh $REMOTE_BUILD_MACHINE -p "$REMOTE_SSH_PORT" \
+	"set -xe && \
+	export ANDROID_HOME=/home/kokeroulis/Android/Sdk && \
+	cd ~ && \
+	mkdir -p $PROJECT_DIR_NAME && \
+	cd $PROJECT_DIR_NAME && \
+	$BUILD_COMMAND"
+fi
 
 # Clean local build dirs.
-rm -rf "$PROJECT_DIR"/build "$PROJECT_DIR"/*/build
-mkdir -p "$PROJECT_DIR/build/"
+if [ $USE_RSYNC == false ]; then
+	rm -rf "$PROJECT_DIR"/build "$PROJECT_DIR"/*/build
+	mkdir -p "$PROJECT_DIR/build/"
+fi
 
 # Copy build results from remote machine to local.
-scp -P "$REMOTE_SSH_PORT" "$REMOTE_BUILD_MACHINE":~/"$PROJECT_DIR_NAME"/remotely_built_project.tar "$PROJECT_DIR/build/"
-
-# Unarchive build results.
-pushd "$PROJECT_DIR"
-eval "$LOCAL_UNARCHIVE_COMMAND"
-popd
+if [ $USE_RSYNC == false ]; then
+	scp -P "$REMOTE_SSH_PORT" "$REMOTE_BUILD_MACHINE":~/"$PROJECT_DIR_NAME"/remotely_built_project.tar "$PROJECT_DIR/build/"
+	# Unarchive build results.
+	pushd "$PROJECT_DIR"
+	eval "$LOCAL_UNARCHIVE_COMMAND"
+	popd
+else
+	eval $RSYNC_FROM_REMOTE
+fi
 
 BUILD_END_TIME=`date +%s`
 echo "End time: $( date )"
